+++ R/crew_controller_group.R
@@ -138,14 +138,7 @@ crew_class_controller_group <- R6::R6Class(
         if (scale) {
           scale(throttle = throttle, controllers = controllers)
         }
-        for (controller in control) {
-          if (controller$unpopped() > 0L) {
-            envir$result <- TRUE
-            return(TRUE)
-          }
-        }
-        .relay$wait()
-        FALSE
+        envir$result <- .relay$wait()
       }
       crew_retry(
         fun = iterate,
@@ -163,7 +156,7 @@ crew_class_controller_group <- R6::R6Class(
       scale,
       throttle
     ) {
-      if (sum(map_int(control, ~ length(.x$tasks))) < 1L) {
+      if (size(controllers = controllers) < 1L) {
         return(TRUE)
       }
       envir <- new.env(parent = emptyenv())
@@ -172,10 +165,10 @@ crew_class_controller_group <- R6::R6Class(
         if (scale) {
           scale(throttle = throttle, controllers = controllers)
         }
-        envir$result <- sum(map_int(control, ~ .x$unresolved())) < 1L
-        if (!envir$result) {
+        if (unresolved() > 0L) {
           .relay$wait()
         }
+        envir$result <- unresolved() < 1L
         envir$result
       }
       crew_retry(
@@ -251,6 +244,14 @@ crew_class_controller_group <- R6::R6Class(
       .throttle$validate()
       invisible()
     },
+    #' @description Number of tasks in the selected controllers.
+    #' @return Non-negative integer, number of tasks in the controller.
+    #' @param controllers Character vector of controller names.
+    #'   Set to `NULL` to select all controllers.
+    size = function(controllers = NULL) {
+      control <- .select_controllers(controllers)
+      sum(map_int(control, ~ .x$size()))
+    },
     #' @description See if the controllers are empty.
     #' @details A controller is empty if it has no running tasks
     #'   or completed tasks waiting to be retrieved with `push()`.
@@ -294,15 +295,6 @@ crew_class_controller_group <- R6::R6Class(
       control <- .select_controllers(controllers)
       sum(map_int(control, ~ .x$unresolved()))
     },
-    #' @description Number of resolved `mirai()` tasks available via `pop()`.
-    #' @return Non-negative integer of length 1,
-    #'   number of resolved `mirai()` tasks available via `pop()`.
-    #' @param controllers Character vector of controller names.
-    #'   Set to `NULL` to select all controllers.
-    unpopped = function(controllers = NULL) {
-      control <- .select_controllers(controllers)
-      sum(map_int(control, ~ .x$unpopped()))
-    },
     #' @description Check if a controller is saturated.
     #' @details A controller is saturated if the number of unresolved tasks
     #'   is greater than or equal to the maximum number of workers.
@@ -328,6 +320,7 @@ crew_class_controller_group <- R6::R6Class(
     start = function(controllers = NULL) {
       control <- .select_controllers(controllers)
       lapply(control, function(controller) controller$start())
+      invisible()
     },
     #' @description Check whether all the given controllers are started.
     #' @details Actually checks whether all the given clients are started.
@@ -839,85 +832,6 @@ crew_class_controller_group <- R6::R6Class(
       )
       out <- tibble::new_tibble(data.table::rbindlist(out, use.names = FALSE))
       if_any(nrow(out), out, NULL)
-    },
-    #' @description Create a `promises::promise()` object to asynchronously
-    #'   pop or collect one or more tasks.
-    #' @details Please be aware that `pop()` or `collect()` will happen
-    #'   asynchronously at a some unpredictable time after the promise object
-    #'   is created, even if your local R process appears to be doing
-    #'   something completely different. This behavior is highly desirable
-    #'   in a Shiny reactive context, but please be careful as it may be
-    #'   surprising in other situations.
-    #' @return A `promises::promise()` object whose eventual value will
-    #'   be a `tibble` with results from one or more popped tasks.
-    #'   If `mode = "one"`, only one task is popped and returned (one row).
-    #'   If `mode = "all"`, then all the tasks are returned in a `tibble`
-    #'   with one row per task (or `NULL` is returned if there are no
-    #'   tasks to pop).
-    #' @param mode Character of length 1, what kind of promise to create.
-    #'   `mode` must be `"one"` or `"all"`. Details:
-    #'   * If `mode` is `"one"`, then the promise is fulfilled (or rejected)
-    #'     when at least one task is resolved and available to `pop()`.
-    #'     When that happens, `pop()` runs asynchronously, pops a result off
-    #'     the task list, and returns a value.
-    #'     If the task succeeded, then the promise
-    #'     is fulfilled and its value is the result of `pop()` (a one-row
-    #'     `tibble` with the result and metadata). If the task threw an error,
-    #'     the error message of the task is forwarded to any error callbacks
-    #'     registered with the promise.
-    #'   * If `mode` is `"all"`, then the promise is fulfilled (or rejected)
-    #'     when there are no unresolved tasks left in the controller.
-    #'     (Be careful: this condition is trivially met in the moment
-    #'     if the controller is empty and you have not submitted any tasks,
-    #'     so it is best to create this kind of promise only after you
-    #'     submit tasks.)
-    #'     When there are no unresolved tasks left,
-    #'     `collect()` runs asynchronously, pops all available results
-    #'     off the task list, and returns a value.
-    #'     If the task succeeded, then the promise
-    #'     is fulfilled and its value is the result of `collect()`
-    #'     (a `tibble` with one row per task result). If any of the tasks
-    #'     threw an error, then the first error message detected is forwarded
-    #'     to any error callbacks registered with the promise.
-    #' @param seconds_interval Positive numeric of length 1, delay in the
-    #'   `later::later()` polling interval to asynchronously check if
-    #'   the promise can be resolved.
-    #' @param scale Deprecated on 2024-04-10 (version 0.9.1.9003)
-    #'   and no longer used. Now, `promise()` always turns on auto-scaling
-    #'   in a private `later` loop (if not already activated).
-    #' @param throttle Deprecated on 2024-04-10 (version 0.9.1.9003)
-    #'   and no longer used. Now, `promise()` always turns on auto-scaling
-    #'   in a private `later` loop (if not already activated).
-    #' @param controllers Not used. Included to ensure the signature is
-    #'   compatible with the analogous method of controller groups.
-    promise = function(
-      mode = "one",
-      seconds_interval = 0.1,
-      scale = NULL,
-      throttle = NULL,
-      controllers = NULL
-    ) {
-      # Tested in tests/interactive/test-promises.R.
-      # nocov start
-      crew_deprecate(
-        name = "controller$promise()",
-        date = "2024-04-19",
-        version = "> 0.9.2",
-        alternative = paste(
-          "see https://wlandau.github.io/crew/articles/shiny.html for the",
-          "latest on using {crew} in promise-driven Shiny apps"
-        ),
-        value = TRUE,
-        frequency = "once"
-      )
-      autoscale(controllers = controllers)
-      controller_promise(
-        controller = self,
-        mode = mode,
-        seconds_interval = seconds_interval,
-        controllers = controllers
-      )
-      # nocov end
     },
     #' @description Wait for tasks.
     #' @details The `wait()` method blocks the calling R session and