+++ tests/testthat/test-crew_controller_group.R
@@ -43,6 +43,7 @@ crew_test("crew_controller_group()", {
   expect_equal(length(x$pids()), 1L)
   x$start()
   expect_true(x$wait(mode = "all", seconds_timeout = 30))
+  expect_equal(x$size(), 0L)
   expect_true(x$started())
   expect_true(x$empty())
   expect_false(x$saturated())
@@ -69,7 +70,6 @@ crew_test("crew_controller_group()", {
     sort(
       c(
         "controller",
-        "tasks",
         "seconds",
         "success",
         "error",
@@ -86,6 +86,7 @@ crew_test("crew_controller_group()", {
     name = "task_pid",
     controller = "b"
   )
+  expect_equal(x$size(), 1L)
   # covr on GitHub Actions mysteriously causes problems
   # in many crew tests.
   covr_ci <- isTRUE(as.logical(Sys.getenv("R_COVR", "false"))) &&
@@ -540,7 +541,6 @@ crew_test("controller group map() works", {
   expect_true(is.character(out$worker))
   expect_equal(out$controller, rep(a$launcher$name, 2L))
   sum <- x$summary()
-  expect_equal(sum$tasks, 2L)
   expect_equal(sum$success, 2L)
   expect_equal(sum$error, 0L)
   expect_equal(sum$warning, 0L)
@@ -893,9 +893,9 @@ crew_test("crash detection with backup controllers in a group", {
     crew_retry(
       ~ {
         x$scale()
-        isTRUE(a$launcher$instances$online) ||
-          isTRUE(b$launcher$instances$online) ||
-          isTRUE(c$launcher$instances$online)
+        isTRUE(a$client$status()["connections"] > 0L) ||
+          isTRUE(b$client$status()["connections"] > 0L) ||
+          isTRUE(c$client$status()["connections"] > 0L)
       },
       seconds_interval = 0.1,
       seconds_timeout = 60